use channels for comms
this also refactors the app setup and messaging modules
