+++ .cargo/config.toml
@@ -1,0 +1,3 @@
+[build]
+target = "wasm32-unknown-unknown"
+rustflags = '-Ctarget-feature=+atomics,+bulk-memory,+mutable-globals'
