+++ .gitignore
@@ -42,4 +42,3 @@ target/
 /wasm-pack.log
 .vscode/
 .idea/
-.cargo
