+++ src/bin/app.rs
@@ -1,0 +1,3 @@
+fn main() {
+    println!("hello world!");
+}
