+++ src/platform/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(target_arch = "wasm32")]
+pub mod web;
+
+#[cfg(not(target_arch = "wasm32"))]
+mod native;
