+++ src/platform/web/worker/workerGen.js
