+++ static/index.html
@@ -14,8 +14,8 @@
       height: 100%;
     }
     canvas {
-      width: 90%;
-      height: 300px;
+      width: 100%;
+      height: 100vh;
       background-color: yellowgreen;
     }
   </style>
@@ -26,7 +26,6 @@
       <ul>
       </ul>
     </nav>
-    <h1 class="title">Enjoy WebGPU + WASM + Webpack with 🍰☕</h1>
   </header>
   <main>
     <section>
