+++ vite.config.js
@@ -16,6 +16,11 @@ export default defineConfig({
   worker: {
     format: "es",
   },
+  resolve: {
+    alias: {
+      'pkg': '/home/akash/workspace/yawn/pkg'
+    }
+  },
   plugins: [
     // Makes us be able to use top level await for wasm.
     // Otherwise, we can restrict build.target to 'es2022', which allows top level await.
@@ -23,8 +28,16 @@ export default defineConfig({
   ],
   server: {
     port: 8080,
+    headers: {
+      'Cross-Origin-Embedder-Policy': 'require-corp',
+      'Cross-Origin-Opener-Policy': 'same-origin',
+    },
   },
   preview: {
     port: 8080,
+    headers: {
+      'Cross-Origin-Embedder-Policy': 'require-corp',
+      'Cross-Origin-Opener-Policy': 'same-origin',
+    },
   },
 });
